Fit Bayesian generalised (non-)linear multilevel compositional model via full Bayesian inference using brms,

This function fits a brm model with multilevel ILR coordinates.

Usage

brmcoda(formula, compilr, ...)

Arguments

formula

A object of class formula, brmsformula: A symbolic description of the model to be fitted. Details of the model specification can be found in brmsformula.

compilr

A compilr object containing data of composition, ILR coordinates, and other variables used in the model.

...

Further arguments passed to brm.

Value

A list with two elements

CompIlr An object of class compilr used in the brm model.

Model An object of class brmsfit, which contains the posterior draws along with many other useful information about the model.

Examples

if (FALSE) {
data(mcompd)
data(sbp)
cilr <- compilr(data = mcompd, sbp = sbp, parts = c("TST", "WAKE", "MVPA", "LPA", "SB"), idvar = "ID")

# inspects ILRs before passing to brmcoda
names(cilr$BetweenILR)
names(cilr$WithinILR)
names(cilr$TotalILR)

# model with compositional predictor at between and within-person levels
m1 <- brmcoda(compilr = cilr, 
             formula = STRESS ~ bilr1 + bilr2 + bilr3 + bilr4 + wilr1 + wilr2 + wilr3 + wilr4 + (1 | ID), 
             chain = 1, iter = 500)

# model with compositional outcome  
m2 <- brmcoda(compilr = cilr, 
              formula = mvbind(ilr1, ilr2, ilr3, ilr4) ~ STRESS + Female + (1 | ID),
             chain = 1, iter = 500) 
               
}