Estimating Average Marginal Effects for Between-person Isotemporal Substitution Model.

Using a fitted model object, estimates the average marginal difference in outcomes when compositional variables are substituted for a specific period at between-person level. The resulting bsubmargins encapsulates substitution estimation across all compositional variables present in the brmcoda object.

Usage

bsubmargins(object, substitute, minute = 60L, ...)

Arguments

object

A brmcoda object. Required.

substitute

A data.frame or data.table of the possible substitution of variables. This dataset can be computed using function possub. Required.

minute

A integer or numeric value indicating the maximum minute for which substitution model is desired. Default to 60L (i.e., the model loops through 1:60L minutes).

...

Additional arguments to be passed to describe_posterior.

Value

A list containing the result of isotemporal multilevel substitution model. Each elements of the list is the substitution estimation for a compositional variables, which include at least six core items.

Mean Posterior means.

CI_low and

CI_high 95% credible intervals.

MinSubstituted Minute substituted within the composition.

SubstituteCompositional variables to be substituted from/to.

PredictorCentral compositional variable to be substituted from/to.

Examples

data(psub)
if (FALSE) {
cilr <- compilr(data = mcompd, sbp = sbp, 
                parts = c("TST", "WAKE", "MVPA", "LPA", "SB"), idvar = "ID")

m <- brmcoda(compilr = cilr, 
             formula = STRESS ~ bilr1 + bilr2 + bilr3 + bilr4 + wilr1 + 
             wilr2 + wilr3 + wilr4 + Female + (1 | ID))
               
testbs <- bsub(object = m, substitute = psub, minute = 5)

library(doFuture)
registerDoFuture()
plan(multisession, workers = 5)
system.time(testbsm <- bsubmargins(object = m, substitute = psub, minute = 5))
}