Compute Between-person, Within-person, and Total Composition and Isometric log ratio transform of a (dataset of) composition(s)

Computes sets of compositions and IRLs for Multilevel Compositional Data models.

Usage

compilr(data, sbp, parts, total = 1440, idvar = "ID")

Arguments

data

A data.frame or data.table containing data of all variables used in the analysis. It must include a composition and a ID variable. Required.

sbp

A signary matrix indicating sequential binary partition. Required.

parts

A character vector specifying the names of compositional variables to be used. Required.

total

A numeric value of the total amount to which the compositions should be closed. Default to 1440.

idvar

A character string specifying the name of the variable containing IDs. Default to ID.

Value

A list with twelve elements.

BetweenComp A vector of class acomp representing one closed between-person composition or a matrix of class acomp representing multiple closed between-person compositions each in one row.

WithinComp A vector of class acomp representing one closed within-person composition or a matrix of class acomp representing multiple closed within-person compositions each in one row.

TotalComp A vector of class acomp representing one closed total composition or a matrix of class acomp representing multiple closed total compositions each in one row.

BetweenILR Isometric log ratio transform of between-person composition.

WithinILR Isometric log ratio transform of within-person composition.

TotalILR Isometric log ratio transform of total composition.

data The user's dataset or imputed dataset if the input data contains zeros.

psi A ILR matrix associated with user-defined partition structure.

sbp The user-defined sequential binary partition matrix.

parts Names of compositional variables.

idvar Name of the variable containing IDs.

total Total amount to which the compositions is closed.

Examples

## Example 1 - Dataset with no 0
data(mcompd)
data(sbp)
cilr1 <- compilr(data = mcompd, sbp = sbp, 
                 parts = c("TST", "WAKE", "MVPA", "LPA", "SB"), idvar = "ID")

## Example 2 - Dataset with 0s
## Impute a 0 in 'mcompd'
mcompd[3, 1] <- 0
cilr2 <- compilr(data = mcompd, sbp = sbp, 
                 parts = c("TST", "WAKE", "MVPA", "LPA", "SB"), idvar = "ID")
#> This dataset of composition contains zero(s);
#>   It is now imputed using the Log-ratio EM algorithm.
#>   For more details, please see ?zCompositions::lrEM
#>   If you would like to deal with zeros using other methods,
#>   please do so before running 'compilr'.
#> No. iterations to converge:  2 
#> 

str(cilr1)
#> List of 12
#>  $ BetweenComp: 'acomp' num [1:3540, 1:5] 0.3432 0.3432 0.3432 0.3432 0.3432 0.0501 0.0501 0.0501 0.0501 0.0501 0.1731 0.1731 0.1731 0.1731 0.1731 0.3499 0| __truncated__ ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "BTST" "BWAKE" "BMVPA" "BLPA" ...
#>  $ WithinComp : 'acomp' num [1:3540, 1:5] 0.2231 0.1962 0.1234 0.195 0.2578 0.2787 0.1449 0.1284 0.192 0.2333 0.2426 0.0996 0.4416 0.1903 0.0878 0.1856 0.2| __truncated__ ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "WTST" "WWAKE" "WMVPA" "WLPA" ...
#>  $ TotalComp  : 'acomp' num [1:3540, 1:5] 0.3766 0.3181 0.1879 0.3649 0.452 0.0687 0.0343 0.0286 0.0525 0.0598 0.2066 0.0814 0.3394 0.1796 0.0776 0.3194 0.| __truncated__ ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "TST" "WAKE" "MVPA" "LPA" ...
#>  $ BetweenILR : 'rmult' num [1:3540, 1:4] -0.684 -0.684 -0.684 -0.684 -0.684 0.747 0.747 0.747 0.747 0.747 -1.299 -1.299 -1.299 -1.299 -1.299 -0.498 -0.498| __truncated__ ...
#>   ..- attr(*, "orig")= 'acomp' num [1:3540, 1:5] 0.3432 0.3432 0.3432 0.3432 0.3432 0.0501 0.0501 0.0501 0.0501 0.0501 0.1731 0.1731 0.1731 0.1731 0.1731 0.3499 0| __truncated__ ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:5] "TST" "WAKE" "MVPA" "LPA" ...
#>   ..- attr(*, "V")= num [1:5, 1:4] -0.224 -0.224 -0.224 -0.224 0.894 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:4] "bilr1" "bilr2" "bilr3" "bilr4"
#>  $ WithinILR  : 'rmult' num [1:3540, 1:4] -1.0655 0.4788 -1.8293 0.392 0.3351 -0.0353 0.1866 -0.6156 0.0974 0.4607 0.2225 -0.1072 -0.8221 0.1184 0.5294 0.1| __truncated__ ...
#>   ..- attr(*, "orig")= 'acomp' num [1:3540, 1:5] 0.2231 0.1962 0.1234 0.195 0.2578 0.2787 0.1449 0.1284 0.192 0.2333 0.2426 0.0996 0.4416 0.1903 0.0878 0.1856 0.2| __truncated__ ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:5] "TST" "WAKE" "MVPA" "LPA" ...
#>   ..- attr(*, "V")= num [1:5, 1:4] -0.224 -0.224 -0.224 -0.224 0.894 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:4] "wilr1" "wilr2" "wilr3" "wilr4"
#>  $ TotalILR   : 'rmult' num [1:3540, 1:4] -1.749 -0.205 -2.513 -0.292 -0.349 0.712 0.934 0.132 0.845 1.208 -1.077 -1.406 -2.121 -1.181 -0.77 -0.308 -1.214 | __truncated__ ...
#>   ..- attr(*, "orig")= 'acomp' num [1:3540, 1:5] 0.3766 0.3181 0.1879 0.3649 0.452 0.0687 0.0343 0.0286 0.0525 0.0598 0.2066 0.0814 0.3394 0.1796 0.0776 0.3194 0.| __truncated__ ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:5] "TST" "WAKE" "MVPA" "LPA" ...
#>   ..- attr(*, "V")= num [1:5, 1:4] -0.224 -0.224 -0.224 -0.224 0.894 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:4] "ilr1" "ilr2" "ilr3" "ilr4"
#>  $ data       :Classes ‘data.table’ and 'data.frame':	3540 obs. of  9 variables:
#>   ..$ TST   : num [1:3540] 494 494 494 494 494 ...
#>   ..$ WAKE  : num [1:3540] 72.2 72.2 72.2 72.2 72.2 ...
#>   ..$ MVPA  : num [1:3540] 249 249 249 249 249 ...
#>   ..$ LPA   : num [1:3540] 504 504 504 504 504 ...
#>   ..$ SB    : num [1:3540] 120 120 120 120 120 ...
#>   ..$ ID    : int [1:3540] 185 185 185 185 185 185 185 185 185 185 ...
#>   ..$ Age   : num [1:3540] 29.7 29.7 29.7 29.7 29.7 ...
#>   ..$ Female: num [1:3540] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ STRESS: num [1:3540] 3.67 7.21 2.84 1.51 7.75 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ psi        : num [1:5, 1:4] -0.224 -0.224 -0.224 -0.224 0.894 ...
#>  $ sbp        : num [1:4, 1:5] -1 1 0 0 -1 -1 1 0 -1 -1 ...
#>  $ parts      : chr [1:5] "TST" "WAKE" "MVPA" "LPA" ...
#>  $ idvar      : chr "ID"
#>  $ total      : num 1440
#>  - attr(*, "class")= chr "compilr"

## cleanup
rm(cilr1, cilr2, mcompd, sbp)
#> Warning: object 'sbp' not found