A series of functions for single level multivariate compositional models

Fit Bayesian generalised (non-)linear multivariate compositional model via full Bayesian inference using Stan, when composition is the outcome.

Usage

mvcoda(data, formula, sbp, parts, ...)

Arguments

data

A data.frame or data.table containing data of all variables used in the analysis. It must include a composition and a ID variable. Required.

formula

A object of class brmsformula, mvbrmsformula: A symbolic description of the model to be fitted.

sbp

A signary matrix indicating sequential binary partition. Required. Details of the model specification can be found in mvbrmsformula.

parts

A character vector specifying the names of compositional variables. Required.

...

Further arguments passed to brm.

Value

A list of results. TODO

Examples

data(mcompd)
data(sbp)
## run mvcoda
mv <- mvcoda(mcompd, mvbind(ilr1, ilr2, ilr3, ilr4) ~ STRESS + Age + Female,
              sbp, c("TST", "WAKE", "MVPA", "LPA", "SB"))
#> Setting 'rescor' to TRUE by default for this model
#> Warning: In the future, 'rescor' will be set to FALSE by default for all models. It is thus recommended to explicitely set 'rescor' via 'set_rescor' instead of using the default.
#> Compiling Stan program...
#> Start sampling
#> 
#> SAMPLING FOR MODEL 'dda3c48844ac42a80ce1fbb4c1f43ccc' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.002871 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 28.71 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 45.9234 seconds (Warm-up)
#> Chain 1:                33.9804 seconds (Sampling)
#> Chain 1:                79.9038 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dda3c48844ac42a80ce1fbb4c1f43ccc' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.00217 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 21.7 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 44.2952 seconds (Warm-up)
#> Chain 2:                32.5362 seconds (Sampling)
#> Chain 2:                76.8313 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dda3c48844ac42a80ce1fbb4c1f43ccc' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.002136 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 21.36 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 45.7632 seconds (Warm-up)
#> Chain 3:                29.5816 seconds (Sampling)
#> Chain 3:                75.3448 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dda3c48844ac42a80ce1fbb4c1f43ccc' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.002108 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 21.08 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 63.9652 seconds (Warm-up)
#> Chain 4:                31.4024 seconds (Sampling)
#> Chain 4:                95.3675 seconds (Total)
#> Chain 4: