Draws from the Posterior Predictive Distribution

Compute posterior draws of the posterior predictive distribution of a brmsfit model in the brmcoda object. Can be performed for the data used to fit the model (posterior predictive checks) or for new data. By definition, these draws have higher variance than draws of the expected value of the posterior predictive distribution computed by fitted.brmcoda. This is because the residual error is incorporated in posterior_predict. However, the estimated means of both methods averaged across draws should be very similar.

Usage

# S3 method for brmcoda
predict(
  object,
  newdata = NULL,
  re_formula = NULL,
  scale = c("response", "linear", "comp"),
  summary = TRUE,
  ...
)

Arguments

object

An object of class brmcoda.

newdata

An optional data.frame for which to evaluate predictions. If NULL (default), the original data of the model is used. NA values within factors are interpreted as if all dummy variables of this factor are zero. This allows, for instance, to make predictions of the grand mean when using sum coding.

re_formula

formula containing group-level effects to be considered in the prediction. If NULL (default), include all group-level effects; if NA, include no group-level effects.

scale

Either "response" or "linear" or "comp". If either "response" or "linear", results are returned using the corresponding method in predict.brmsfit. "comp" is only relevant to multivariate models of class mvbrmsformula (i.e., when response is composition). If "comp", results are returned on the compositional scale of the response variable.

summary

Should summary statistics be returned instead of the raw values? Default is TRUE.

...

Further arguments passed to predict.brmsfit that control additional aspects of prediction.

Value

An array of predicted response values. If summary = FALSE the output resembles those of

posterior_predict.brmsfit.

If summary = TRUE the output depends on the family: For categorical and ordinal families, the output is an N x C matrix, where N is the number of observations, C is the number of categories, and the values are predicted category probabilities. For all other families, the output is a N x E matrix where E = 2 + length(probs) is the number of summary statistics: The Estimate column contains point estimates (either mean or median depending on argument robust), while the

Est.Error column contains uncertainty estimates (either standard deviation or median absolute deviation depending on argument

robust). The remaining columns starting with Q contain quantile estimates as specified via argument probs.

See also