Multilevel Compositional Substitution Model

Estimate the difference in an outcome when compositional parts are substituted for specific unit(s). The substitution output encapsulates the substitution results for all compositional parts present in the brmcoda object.

Usage

substitution(
  object,
  delta,
  basesub = NULL,
  summary = TRUE,
  ref = c("grandmean", "clustermean"),
  level = c("between", "within"),
  weight = c("equal", "proportional"),
  ...
)

Arguments

object

A fitted brmcoda object.

delta

A integer, numeric value or vector indicating the amount of substituted change between compositional parts.

basesub

A data.frame or data.table of the base possible substitution of compositional parts. This data set can be computed using function basesub. If NULL, all possible pairwise substitution of compositional parts are used.

summary

A logical value. Should the estimate at each level of the reference grid (FALSE) or their average (TRUE) be returned? Default is TRUE. Only applicable for model with covariates in addition to the isometric log-ratio coordinates (i.e., adjusted model).

ref

Either a character value or vector or a dataset. Can be "grandmean" and/or "clustermean", or a data.frame or data.table of user's specified reference grid consisting of combinations of covariates over which predictions are made. User's specified reference grid is only possible for simple substitution.

level

A character string or vector. Should the estimate be at the "between" and/or "within" level?

weight

A character value specifying the weight to use in calculation of the reference composition. If "equal", give equal weight to units (e.g., individuals). If "proportional", weights in proportion to the frequencies of units being averaged (e.g., observations across individuals) Default is equal.

...

Additional arguments passed to describe_posterior.

Value

A list containing the results of multilevel compositional substitution model. The first four lists contain the results of the substitution estimation for a compositional part.

Mean Posterior means.

CI_low and CI_high 95% credible intervals.

Delta Amount substituted across compositional parts.

From Compositional part that is substituted from.

To Compositional parts that is substituted to.

Level Level where changes in composition takes place. Either between or within.

Reference Either grandmean, clustermean, or users.

Examples

# \donttest{
if(requireNamespace("cmdstanr")){
  cilr <- compilr(data = mcompd, sbp = sbp,
                  parts = c("TST", "WAKE", "MVPA", "LPA", "SB"),
                  idvar = "ID", total = 1440)
  
  # model with compositional predictor at between and between-person levels
  m <- brmcoda(compilr = cilr,
               formula = Stress ~ bilr1 + bilr2 + bilr3 + bilr4 +
                 wilr1 + wilr2 + wilr3 + wilr4 + (1 | ID),
               chain = 1, iter = 500, backend = "cmdstanr")
  
  subm <- substitution(object = m, delta = 5)
}# }
#> Error: CmdStan path has not been set yet. See ?set_cmdstan_path.