Summary substitution

This method allows for summarising an existing substitution object.

Usage

# S3 method for substitution
summary(object, delta, specs, ref, level, digits = 2, ...)

Arguments

object

A substitution object containing the output of substitution models.

delta

A integer, numeric value or vector indicating the desired delta at which substitution results should be summarised. Default to all delta available in substitution object.

ref

Either a character value or vector ((grandmean and/or unitmean or users), Default to all ref available in substitution object .

level

A character string or vector (between and/or within). Default to all level available in substitution object .

digits

A integer value used for number formatting. Default to 3.

...

generic argument, not in use.

Value

A summary of substitution object with twelve elements.

BetweenComp A vector of class acomp representing one closed between-person composition or a matrix of class acomp representing multiple closed between-person compositions each in one row.

WithinComp A vector of class acomp representing one closed within-person composition or a matrix of class acomp representing multiple closed within-person compositions each in one row.

TotalComp A vector of class acomp representing one closed total composition or a matrix of class acomp representing multiple closed total compositions each in one row.

BetweenILR Isometric log ratio transform of between-person composition.

WithinILR Isometric log ratio transform of within-person composition.

TotalILR Isometric log ratio transform of total composition.

data The user's dataset or imputed dataset if the input data contains zeros.

psi A ILR matrix associated with user-defined partition structure.

sbp The user-defined sequential binary partition matrix.

parts Names of compositional variables.

idvar Name of the variable containing IDs.

total Total amount to which the compositions is closed.