Indices from a (dataset of) Multilevel Composition(s)

Compute sets of compositions and IRLs for multilevel compositional data

Usage

compilr(data, sbp, parts, total = 1, idvar = "ID")

Arguments

data

A data.frame or data.table containing data of all variables used in the analysis. It must include a composition and a ID variable. Required.

sbp

A signary matrix indicating sequential binary partition. Required.

parts

A character vector specifying the names of compositional variables to be used.

total

A numeric value of the total amount to which the compositions should be closed. Default is 1.

idvar

A character string specifying the name of the variable containing IDs. Default is "ID".

Value

A compilr object with twelve elements.

BetweenComp A vector of class acomp representing one closed between-person composition or a matrix of class acomp representing multiple closed between-person compositions each in one row.

WithinComp A vector of class acomp representing one closed within-person composition or a matrix of class acomp representing multiple closed within-person compositions each in one row.

TotalComp A vector of class acomp representing one closed total composition or a matrix of class acomp representing multiple closed total compositions each in one row.

BetweenILR Isometric log ratio transform of between-person composition.

WithinILR Isometric log ratio transform of within-person composition.

TotalILR Isometric log ratio transform of total composition.

data The user's dataset or imputed dataset if the input data contains zeros.

psi A ILR matrix associated with user-defined partition structure.

sbp The user-defined sequential binary partition matrix.

parts Names of compositional variables.

idvar Name of the variable containing IDs.

total Total amount to which the compositions is closed.

Examples

cilr <- compilr(data = mcompd, sbp = sbp,
                parts = c("TST", "WAKE", "MVPA", "LPA", "SB"),
                idvar = "ID", total = 1440)
str(cilr)
#> List of 12
#>  $ BetweenComp: 'acomp' num [1:3540, 1:5] 494.2 494.2 494.2 494.2 494.2 72.2 72.2 72.2 72.2 72.2 249.3 249.3 249.3 249.3 249.3 503.8 503.8 503.8 503.8 503.| __truncated__ ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "BTST" "BWAKE" "BMVPA" "BLPA" ...
#>  $ WithinComp : 'acomp' num [1:3540, 1:5] 0.2231 0.1962 0.1234 0.195 0.2578 0.2787 0.1449 0.1284 0.192 0.2333 0.2426 0.0996 0.4416 0.1903 0.0878 0.1856 0.2| __truncated__ ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "WTST" "WWAKE" "WMVPA" "WLPA" ...
#>  $ TotalComp  : 'acomp' num [1:3540, 1:5] 542.3 458.1 270.6 525.4 650.8 99 49.4 41.1 75.6 86.1 297.4 117.3 488.7 258.7 111.8 459.9 652.9 625 397.9 436.5 41| __truncated__ ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "TST" "WAKE" "MVPA" "LPA" ...
#>  $ BetweenILR : 'rmult' num [1:3540, 1:4] -0.29534 -0.29534 -0.29534 -0.29534 -0.29534 1.36014 1.36014 1.36014 1.36014 1.36014 0.00961 0.00961 0.00961 0.00| __truncated__ ...
#>   ..- attr(*, "orig")= 'acomp' num [1:3540, 1:5] 0.3432 0.3432 0.3432 0.3432 0.3432 0.0501 0.0501 0.0501 0.0501 0.0501 0.1731 0.1731 0.1731 0.1731 0.1731 0.3499 0| __truncated__ ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:5] "TST" "WAKE" "MVPA" "LPA" ...
#>   ..- attr(*, "V")= num [1:5, 1:4] 0.548 0.548 -0.365 -0.365 -0.365 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:4] "bilr1" "bilr2" "bilr3" "bilr4"
#>  $ WithinILR  : 'rmult' num [1:3540, 1:4] 0.5822 -0.177 -0.1904 -0.0204 0.5006 -0.1573 0.2142 -0.028 0.0112 0.0705 0.6174 -0.8432 1.3248 -0.0428 -0.6989 0.| __truncated__ ...
#>   ..- attr(*, "orig")= 'acomp' num [1:3540, 1:5] 0.2231 0.1962 0.1234 0.195 0.2578 0.2787 0.1449 0.1284 0.192 0.2333 0.2426 0.0996 0.4416 0.1903 0.0878 0.1856 0.2| __truncated__ ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:5] "TST" "WAKE" "MVPA" "LPA" ...
#>   ..- attr(*, "V")= num [1:5, 1:4] 0.548 0.548 -0.365 -0.365 -0.365 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:4] "wilr1" "wilr2" "wilr3" "wilr4"
#>  $ TotalILR   : 'rmult' num [1:3540, 1:4] 0.2869 -0.4724 -0.4858 -0.3157 0.2053 1.2029 1.5743 1.3321 1.3713 1.4307 0.627 -0.8336 1.3344 -0.0332 -0.6893 1.7| __truncated__ ...
#>   ..- attr(*, "orig")= 'acomp' num [1:3540, 1:5] 0.3766 0.3181 0.1879 0.3649 0.452 0.0687 0.0343 0.0286 0.0525 0.0598 0.2066 0.0814 0.3394 0.1796 0.0776 0.3194 0.| __truncated__ ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:5] "TST" "WAKE" "MVPA" "LPA" ...
#>   ..- attr(*, "V")= num [1:5, 1:4] 0.548 0.548 -0.365 -0.365 -0.365 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:4] "ilr1" "ilr2" "ilr3" "ilr4"
#>  $ data       :Classes ‘data.table’ and 'data.frame':	3540 obs. of  10 variables:
#>   ..$ ID    : int [1:3540] 185 185 185 185 185 185 185 185 185 185 ...
#>   ..$ Time  : int [1:3540] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ Stress: num [1:3540] 3.67 7.21 2.84 1.51 7.75 ...
#>   ..$ TST   : num [1:3540] 542 458 271 525 651 ...
#>   ..$ WAKE  : num [1:3540] 99 49.4 41.1 75.6 86.1 ...
#>   ..$ MVPA  : num [1:3540] 297 117 489 259 112 ...
#>   ..$ LPA   : num [1:3540] 460 653 625 398 436 ...
#>   ..$ SB    : num [1:3540] 41.4 162.3 14.5 182.5 154.8 ...
#>   ..$ Age   : num [1:3540] 29.7 29.7 29.7 29.7 29.7 ...
#>   ..$ Female: num [1:3540] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..- attr(*, "index")= int(0) 
#>   .. ..- attr(*, "__ID")= int [1:3540] 981 982 983 984 985 986 987 988 989 990 ...
#>  $ psi        : num [1:5, 1:4] 0.548 0.548 -0.365 -0.365 -0.365 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:5] "TST" "WAKE" "MVPA" "LPA" ...
#>   .. ..$ : NULL
#>  $ sbp        : num [1:4, 1:5] 1 1 0 0 1 -1 0 0 -1 0 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:5] "TST" "WAKE" "MVPA" "LPA" ...
#>  $ parts      : chr [1:5] "TST" "WAKE" "MVPA" "LPA" ...
#>  $ idvar      : chr "ID"
#>  $ total      : num 1440
#>  - attr(*, "class")= chr "compilr"